backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.7
python-docx==1.1.0
docx2pdf==0.1.4
fpdf==1.7.2
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2